requirements.txt
opencv_python_headless==4.5.2.54
streamlit==0.82.0
mediapipe==0.8.4.2
numpy==1.18.5
Pillow==8.2.0
markupsafe==2.0.1
scikit-learn==1.1.1
